refactoring into more expandable classes
