+++ config.py
@@ -17,18 +17,18 @@ config = {
     # General
     "AGENT": "n-steps A2C",
     "GAMMA": 0.99,
-    "NB_TIMESTEPS_TRAIN": 2e5,
-    "NB_EPISODES_TEST": 10,
+    "NB_TIMESTEPS_TRAIN": 1e4,
+    "NB_EPISODES_TEST": 1,
     "VALUE_FACTOR": 0.5,
     "ENTROPY_FACTOR": 0,
     "KL_FACTOR": 0.0000,
-    "LEARNING_START": 1e5,
+    "LEARNING_START": 1e3,
     # Specific
     "N_STEPS": 1,
     # NETWORKS
     "RECURRENT": False,
-    "GRADIENT_CLIPPING": None,
-    "LEARNING_RATE": 5e-4,
+    "GRADIENT_CLIPPING": 0.5,
+    "LEARNING_RATE": 5e-5,
     "LEARNING_RATE_END": 1e-6,
     # RNN
     "HIDDEN_SIZE": 64,
@@ -48,7 +48,7 @@ config = {
     "N_EXPERIMENTS": 3,
     "EARLY_STOPPING_STEPS": 10000,
     # Logging
-    "logging": "wandb",
+    "logging": "Tensorboard",
     # Normalization
     "SCALING": True,
     "SCALING_METHOD": "standardize",
@@ -58,4 +58,3 @@ config = {
     "BUFFER_SIZE": 1,
     "NORMALIZE_ADVANTAGES": False,
 }
-
